auth_app/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.26.0
aiofiles==23.2.1 